Move PortalProvider to new pkg, updates request func to use async/await
