+++ PortalSwift/Sources/PortalSwift/Provider/PortalProviderDelegate.swift
@@ -1,0 +1,3 @@
+public protocol PortalProviderDelegate {
+  func portalProvider(didRequestSigningApprovalForRequest: PortalProviderRequestWithId) -> Bool
+}
